+++ tests/test_humanitarian_needs.py
@@ -7,12 +7,7 @@ from hapi_schema.db_humanitarian_needs import (
     DBHumanitarianNeeds,
     view_params_humanitarian_needs,
 )
-from hapi_schema.utils.enums import (
-    DisabledMarker,
-    Gender,
-    PopulationGroup,
-    PopulationStatus,
-)
+from hapi_schema.utils.enums import PopulationStatus
 from hapi_schema.views import prepare_hapi_views
 
 
@@ -34,11 +29,10 @@ def test_humanitarian_needs_view(run_view_test):
             view_humanitarian_needs.c.admin1_code == "FOO-001",
             view_humanitarian_needs.c.location_code == "FOO",
             view_humanitarian_needs.c.population_status == "INN",
-            view_humanitarian_needs.c.population_group == "IDP",
+            view_humanitarian_needs.c.category
+            == "Female - Disabled - Baby - IDP",
             view_humanitarian_needs.c.sector_name
             == "Water Sanitation Hygiene",
-            view_humanitarian_needs.c.gender == "f",
-            view_humanitarian_needs.c.disabled_marker == "y",  # noqa: E712
         ),
     )
 
@@ -65,11 +59,8 @@ def base_parameters():
         admin2_ref=1,
         provider_admin1_name="Provincia 01",
         provider_admin2_name="Distrito B",
-        gender=Gender.ALL,
-        age_range="-1-20",
-        disabled_marker=DisabledMarker.ALL,
+        category="",
         sector_code="*",
-        population_group=PopulationGroup.ALL,
         population_status=PopulationStatus.AFFECTED,
         population=1_000_000,
         reference_period_start=datetime(2023, 1, 1),
@@ -100,47 +91,4 @@ def test_population_positive(run_constraints_test, base_parameters):
             DBHumanitarianNeeds(**modified_params),
         ],
         expected_constraint="population_constraint",
-    )
-
-
-def test_minage(run_constraints_test, base_parameters):
-    """Check that the min_age value is positive, and NULL if
-    age_range is all"""
-    modified_params = {**base_parameters, **dict(age_range="5-10", min_age=-1)}
-    run_constraints_test(
-        new_rows=[
-            DBHumanitarianNeeds(**modified_params),
-        ],
-        expected_constraint="min_age_constraint",
-    )
-    modified_params = {**base_parameters, **dict(age_range="all", min_age=10)}
-    run_constraints_test(
-        new_rows=[
-            DBHumanitarianNeeds(**modified_params),
-        ],
-        expected_constraint="min_age_constraint",
-    )
-
-
-def test_maxage(run_constraints_test, base_parameters):
-    """Check that the max_age is > min_age, and NULL when min_age is NULL"""
-    modified_params = {
-        **base_parameters,
-        **dict(age_range="5-10", min_age=5, max_age=4),
-    }
-    run_constraints_test(
-        new_rows=[
-            DBHumanitarianNeeds(**modified_params),
-        ],
-        expected_constraint="max_age_constraint",
-    )
-    modified_params = {
-        **base_parameters,
-        **dict(age_range="unknown", min_age=None, max_age=10),
-    }
-    run_constraints_test(
-        new_rows=[
-            DBHumanitarianNeeds(**modified_params),
-        ],
-        expected_constraint="max_age_constraint",
     )